fix(ui): consolidate deferred project selection
